Move PathCursor
